Fix uptime/offset bug
Keylog inserts to the DB are in epoch ms, but we only get the keyboad data as uptime ms. This caused a subtle bug, because uptime is paused when the device is in deep sleep, which caused an issue with drift because the offset was only calcaulted on instance creation, and instances survive through deep sleep. This fixes the problem by recalculating the offset every time the session starts, which guarantees the correct uptime ms is being used.